+++ apps/examples/wc-vue3/CHANGELOG.md
@@ -1,5 +1,26 @@
 # wc-vue3
 
+## 0.5.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.29
+
+## 0.5.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.28
+
+## 0.5.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-webc@0.5.27
+
 ## 0.5.13
 
 ### Patch Changes