+++ packages/components/pie-text-input/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @justeattakeaway/pie-text-input
 
+## 0.23.6
+
+### Patch Changes
+
+- [Changed] - ensure the placeholder text isn't obscured by opacity in firefox ([#1747](https://github.com/justeattakeaway/pie/pull/1747)) by [@raoufswe](https://github.com/raoufswe)
+
+  [Changed] - updated the readonly background to use the subtle token
+
+## 0.23.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @justeattakeaway/pie-assistive-text@0.7.1
+
 ## 0.23.4
 
 ### Patch Changes